tg_bot_service/requirements.txt
python-telegram-bot[job-queue]==20.7
httpx==0.25.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2